Rename reducer
