+++ src/js/reducers/timer.js
